Homework 5 (strings) updated (110%). Reworking task about guess game in progress
